High resolution timer for timing fast controllers like lqr

+++ Firmware/Src/Zynq/timer_interrupt.c
@@ -10,9 +10,9 @@ unsigned long TIMER1_getSystemTime_Us() {
 }
 
 /* High-resolution timing function for sub-microsecond measurements */
-unsigned long TIMER1_getSystemTime_Cycles(void) {
+unsigned long long TIMER1_getSystemTime_Cycles(void) {
 		XTime_GetTime(&MyTime);
-		return (unsigned long)MyTime;
+		return (unsigned long long)MyTime;
 }
 
 static void Timer_Intr_Handler(void *CallBackRef);
